feat: deprecate `onUnhandledRequest`
