+++ src/middleware/node/types.ts
@@ -9,6 +9,7 @@ import { Logger } from "../../createLogger";
 export type MiddlewareOptions = {
   path?: string;
   log?: Logger;
+  /** @deprecated `onUnhandledRequest()` is deprecated and will be removed in a future release of `@octokit/webhooks` */
   onUnhandledRequest?: (
     request: IncomingMessage,
     response: ServerResponse
